impl-node: implement mode node.passive property values
The node.passive property can actually take true, in and out as values
to mark what ports to make passive. Keep track of what port direction to
make passive.

Set the passive property also on ports and make all port inherit the
property from the node first (based on direction) and then use the new
port.passive property as an override.

Make the link use the passive property from the ports that it links to
check the passive state of the link.