alsa-mixer: fix the re-attach code for the mixer control element
The new helem must be tracked and old helem must be cleared
to make the code work properly. Introduce the pointer to helem
as the private value for melem and add the necessary code.

Also, add a check for the duplicate mixer elements. The duplicate
mixer element invokes the abort check in alsa-lib. Print a warning
instead and handle the exit gracefully.

Link: https://gitlab.freedesktop.org/pulseaudio/pulseaudio/-/commit/d1675df0cdd5565aa6823afefe0795816228d06e
Signed-off-by: Jaroslav Kysela <perex@perex.cz>