impl-node: handle ready callback correctly
Only stop processing the ready callback if we are a driver and not currently
driving the graph.

Streams that use trigger will also emit the ready callback and are not
driving the graph (but also are not a driver) and should therefore be
allowed to continue to resume_node to schedule the peer nodes.

See #3184