context: Also add driver to always_process inactive nodes
Commit 90b11e3c49c061bae26b020b7b8a492617627205 removed inactive
node from the fallback driver, even if it was always_process.

This should not happen, always_process nodes should stay on a driver in
all cases.

JACK nodes are marked like this and need to stay on the driver even when
inactive or else they don't have a transport when inactive.

Fixes #3189