Revert "module-raop: zero uppet timestamp bits"
This reverts commit 21d16b1ad5499233037097a1e70b3bf1d3935c2b.

The change causes the sound of videos to be way out of sync when streaming from Fedora to Sonos.

It seems the issue is a device-specific quirk, and the change cannot be applied universally to all devices,
thus reverting until a better solution is found that does not affect other devices.